+++ fleet-schedule-demo/data-cache/src/main/java/demo/domain/RoadRepository.java
@@ -10,8 +10,8 @@ import org.springframework.data.repository.query.Param;
 
 public interface RoadRepository extends JpaRepository<Road, Long> {
 
-    Page<Road> findByBeginPoint(@Param("beginPoint") int beginPoint, Pageable pageable); //按起点查询道路，下类似
-    Page<Road> findByEndPoint(@Param("endPoint") int endPoint, Pageable pageable);
-    Page<Road> findByBeginPointAndEndPoint(@Param("beginPoint") int beginPoint, @Param("endPoint") int endPoint, Pageable pageable);
+    Page<Road> findByBeginPoint(@Param("beginPoint") Long beginPoint, Pageable pageable); //按起点查询道路，下类似
+    Page<Road> findByEndPoint(@Param("endPoint") Long endPoint, Pageable pageable);
+    Page<Road> findByBeginPointAndEndPoint(@Param("beginPoint") Long beginPoint, @Param("endPoint") Long endPoint, Pageable pageable);
 
 }
